+++ src/index.js
@@ -1,3 +1,6 @@
 import toPercentageBy from './toPercentageBy';
 
 export default toPercentageBy;
+
+// export with default config
+export const toPercentage = toPercentageBy();