+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [HEAD]
 > Unreleased
 
+* chore(export): expose `toPercentage` with default config.
+
 ## [v1.0.0]
 > Mar 27, 2016